Pan Gestures - Changed Pan gesture prediction amount units to millis
Instead of 0.7 the prediction amount to set should be (0.7 * 16)

+++ dali/integration-api/profiling.cpp
@@ -111,7 +111,7 @@ void SetPanGesturePredictionMode( int mode )
   eventProcessor.SetPanGesturePredictionMode(mode);
 }
 
-void SetPanGesturePredictionAmount(float amount)
+void SetPanGesturePredictionAmount(unsigned int amount)
 {
   GestureEventProcessor& eventProcessor = ThreadLocalStorage::Get().GetGestureEventProcessor();
   eventProcessor.SetPanGesturePredictionAmount(amount);
