+++ dali/internal/event/events/pan-gesture-processor.h
@@ -129,9 +129,9 @@ public: // To be called by GestureEventProcessor
   /**
    * @brief Sets the prediction amount of the pan gesture
    *
-   * @param[in] amount The prediction amount, 0.0f being next vsync and each 1.0f on top is another vsync ahead, can be divisions of (0.5f)
+   * @param[in] amount The prediction amount in milliseconds
    */
-  void SetPredictionAmount(float amount);
+  void SetPredictionAmount(unsigned int amount);
 
 private:
